+++ global_var.h
@@ -28,9 +28,9 @@ extern unsigned long FB2;
 extern unsigned int original_CFG2;
 extern unsigned short ADC_FLG;
 extern char hex[];
-extern volatile unsigned short PW1;					//set initial value for PW1 and PW2
-extern volatile unsigned short PW2;
-extern volatile unsigned short PW3;
+extern volatile unsigned int PW1;					//set initial value for PW1 and PW2
+extern volatile unsigned int PW2;
+extern volatile unsigned int PW3;
 extern char* PW_chars;
 extern int POT_COUNT_DOWN;
 extern int dummy_time;
